backend/pyproject.toml
[project]
name = "axiom-backend"
version = "0.1.0"
description = "Axiom Design Engine - Backend API"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    
    # Redis & Caching
    "redis>=5.0.1",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Validation & Settings
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    
    # HTTP Client
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    
    # Storage
    "boto3>=1.34.0",
    
    # Utilities
    "python-dateutil>=2.8.2",
    "orjson>=3.9.10",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    
    # Code Quality
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    
    # Type Stubs
    "types-python-dateutil>=2.8.19",
    "types-passlib>=1.7.7",
    "types-redis>=4.6.0",
    "boto3-stubs>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
